moved lsrro flowsheet
